IPL High Performers Dataset

This dataset contains IPL 2022 players who have achieved significant performance benchmarks, either scoring over 3000 runs or taking more than 90 wickets across IPL seasons. It is intended to be used for analyzing key batting and bowling metrics for strategic purposes, including IPL auction strategies.

Usage

ipl_high_performers

Format

A data frame with filtered IPL player statistics. The key columns in the dataset are:

Name

The full name of the player.

Surname

The surname of the player, extracted from the full name.

Team

The IPL team that the player is a part of.

RunsScored

The total number of runs scored by the player.

Wickets

The total number of wickets taken by the player.

BattingAVG

The player's batting average (runs per innings).

BattingS/R

The player's strike rate (runs per 100 balls faced).

BowlingAVG

The player's bowling average (runs conceded per wicket).

EconomyRate

The player's bowling economy rate (runs conceded per over).

Source

IPL 2022 Player Statistics Dataset

Examples

# Load the data
data("ipl_high_performers", package = "Ipl")

# View the first few rows of the dataset
head(ipl_high_performers)
#> # A tibble: 6 × 40
#>   Name       Team  Url   Type  ValueinCR `Full Name` Born  Age   `National Side`
#>   <chr>      <chr> <chr> <chr>     <dbl> <chr>       <chr> <chr> <chr>          
#> 1 Shikhar D… PBKS  http… Bats…      8.25 Shikhar Dh… Dece… 36 Y… India          
#> 2 Sandeep S… PBKS  http… Bowl…      0.5  Sandeep Sh… May … 28 Y… India          
#> 3 Bhuvneshw… SRH   http… Bowl…      4.2  Bhuvneshwa… Febr… 32 Y… India          
#> 4 Sanju Sam… RR    http… Wick…     14    Sanju Visw… Nove… 27 Y… India          
#> 5 Yuzvendra… RR    http… Bowl…      6.5  Yuzvendra … July… 31 Y… India          
#> 6 Ravichand… RR    http… All-…      5    Ravichandr… Sept… 35 Y… India          
#> # ℹ 31 more variables: `Batting Style` <chr>, Bowling <chr>, Sport <chr>,
#> #   MatchPlayed <dbl>, InningsBatted <dbl>, NotOuts <dbl>, RunsScored <dbl>,
#> #   HighestInnScore <chr>, `100s` <dbl>, `50s` <dbl>, `4s` <dbl>, `6s` <dbl>,
#> #   BattingAVG <dbl>, `BattingS/R` <dbl>, CatchesTaken <dbl>,
#> #   StumpingsMade <dbl>, Ducks <dbl>, `R/O` <dbl>, InningsBowled <dbl>,
#> #   Overs <dbl>, Maidens <dbl>, RunsConceded <dbl>, Wickets <dbl>, Best <chr>,
#> #   `3s` <dbl>, `5s` <dbl>, BowlingAVG <dbl>, EconomyRate <dbl>, `S/R` <dbl>, …

# Filter for top-performing batsmen
top_batsmen <- ipl_high_performers %>%
  filter(RunsScored > 3000)
#> Error in ipl_high_performers %>% filter(RunsScored > 3000): could not find function "%>%"

# Filter for top-performing bowlers
top_bowlers <- ipl_high_performers %>%
  filter(Wickets > 90)
#> Error in ipl_high_performers %>% filter(Wickets > 90): could not find function "%>%"